+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Healthcheck Changelog
 ### Next Release
-* Fix comaptibility issue with FreeBSD
+* Fix compatibility issue with FreeBSD
 
 ### 1.7.1